Apply MethodRules for each hierarchy context level
MethodRules must be applied for every hierarchy level. Each invocation
of the apply method needs the current instance of the hierarchy get
inserted. Otherwise it cannot operate on it and "do it's magic". An
example is the Mockito JUnit rule that can only properly instantiate all
mocks when it gets invoked for every hierarchy level.

Cf. issue #26 for more details.